all the changes

+++ src/controlller/noteController.js
@@ -1,7 +1,9 @@
 const { response } = require("express");
+const mongoose= require("mongoose")
 const noteModel= require("../models/notes");
 
 const createNote = async (req, res)=>{
+    console.log("Incoming request body:", req.body);
     const {title, description}=req.body;
 
     const newNote= new noteModel({
@@ -25,8 +27,24 @@ const createNote = async (req, res)=>{
 const deleteNote=async (req, res)=>{
     const id=req.params.id;
     try {
-        const note= await noteModel.findByIdAndRemove(id);
-        res.status(202).json(note);
+
+        const curr_note = await noteModel.findById(id);
+
+        if (!curr_note) {
+            return res.status(404).json({ message: "Note not found" });
+        }
+
+        // if (curr_note.userId.toString() !== req.userId) {
+        //     return res.status(403).json({ message: "Unauthorized: You can only delete your own notes" });
+        // }
+
+
+
+        await noteModel.findByIdAndRemove(id);
+        res.status(202).json({
+            message: "Note deleted successfully",
+            deletedNote: curr_note // Sending back the deleted note details
+        });
     } catch (error) {
         console.log(error);
         res.status(500).json({message:"Something went wrong"});
@@ -34,7 +52,12 @@ const deleteNote=async (req, res)=>{
 }
 
 const updateNote=async (req, res)=>{
-    const id= req.params.id;
+    let id= req.params.id;
+
+    if (id.startsWith(':')) {
+        id = id.substring(1);
+    }
+    //console.log(id)
     const {title, description}=req.body;
     const newNote={
         title:title,
@@ -47,7 +70,7 @@ const updateNote=async (req, res)=>{
         res.status(200).json(newNote);
     } catch (error) {
         console.log(error);
-        res.status(500).json({message:"Something went wrong"});
+        res.status(500).json({message:"Something went wrong " +error});
     }
 
     
@@ -61,7 +84,7 @@ const getNote=async (req, res)=>{
 
     } catch (error) {
         console.log(error);
-        res.status(500).json({message:"Something went wrong"});
+        res.status(500).json({message:"Something went wrong", error: error.message});
     }
